pace

+++ app/src/main/java/com/kent/lw/brainendurancetrainingmobileapp/ApvtTask.java
@@ -1,7 +1,7 @@
 package com.kent.lw.brainendurancetrainingmobileapp;
 
 public class ApvtTask extends Task{
-    public ApvtTask(int duration, int intervalFrom, int intervalTo, float volumeFrom, float volumeTo, int noise, int resThreshold) {
+    public ApvtTask(int duration, int intervalFrom, int intervalTo, float volumeFrom, float volumeTo, float noise, int resThreshold) {
         super(duration, intervalFrom, intervalTo, volumeFrom, volumeTo, noise, resThreshold);
     }
 }
