+++ app/src/main/java/com/kent/lw/brainendurancetrainingmobileapp/GonogoTask.java
@@ -1,7 +1,7 @@
 package com.kent.lw.brainendurancetrainingmobileapp;
 
 public class GonogoTask extends Task {
-    public GonogoTask(int duration, int intervalFrom, int intervalTo, float volumeFrom, float volumeTo, int noise, int resThreshold) {
+    public GonogoTask(int duration, int intervalFrom, int intervalTo, float volumeFrom, float volumeTo, float noise, int resThreshold) {
         super(duration, intervalFrom, intervalTo, volumeFrom, volumeTo, noise, resThreshold);
     }
 }
